api/requirements.txt
fastapi
uvicorn
transformers>=4.51.0
torch
supabase
python-dotenv
